requirements.debug.txt
-r requirements.txt

django-debug-toolbar==0.9.4

ipdb==0.7
ipython==1.1.0
